Move string.templatelib specific tests to test_string/test_templatlib.py
